requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
ta>=0.11.0
google-genai>=0.8.0
playwright>=1.44.0
python-dotenv>=1.0.0
APScheduler>=3.10.4
pydantic>=2.0.0
